requirements.txt
Django>=4.2
djangorestframework>=3.14.0
psycopg2-binary>=2.9
python-dotenv>=1.0.1
celery>=5.3
redis>=5.0.1
pandas>=2.1.0
numpy>=1.26.0
python-dateutil>=2.9.0
gunicorn
openpyxl

